+++ source/includes/thmc/two_phase_flow/BrooksCoreyWithHysteresisCO2.h
@@ -48,7 +48,7 @@ class BrooksCoreyWithHysteresisCO2 : public TwoPhaseModel<dim> {
     virtual double ShockSpeed() const;
     virtual double ShockHeight() const;
     
-    virtual void Out( size_t phase ) const;
+    virtual void Out( uint32_t phase ) const;
 
   private:
     BrooksCoreyWithHysteresisCO2();
